refactor: ディレクトリ構成変更
Profileディレクトリ内にCSSも移動する, それに伴うパス変更
