+++ src/App.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import styles from './assets/styles/App.module.scss';
 import { Hero } from './components/Hero/Hero';
-import Profile from './components/Profile';
+import { Profile } from './components/Profile/Profile';
 
 const App = () => {
   return (
