+++ src/components/Profile/Profile.module.scss
@@ -1,5 +1,3 @@
-@import './fonts.scss';
-
 .profile {
   font-family: '8bit-12-regular';
   width: 459px;
